requirements.txt
Pillow
matplotlib
numpy
pandas
scikit-learn
